// src/app/gis/models/mapLayerGroup.ts
export enum MapLayerGroup {
  Adjuster = "Adjuster GPS Layers",
  Agency = "Agency Layers",
  Agri = "Agricultural Layers", // Layers really only specific for Agricutlural purposes
  Base = "Base Layers", // Layers every map basically has
  Insured = "Insured Layers",
  PrecisionAg = "Precision Ag Layers"
}