// src/app/gis/models/mapLayerType.ts
export enum MapLayerType {
  HighRisk = "HighRisk",
  BreachedLevees = "BreachedLevees",
  ColumbiaBasin = "ColumbiaBasin",
  RainfallIndex = "RainfallIndex",
  TYield = "TYield",
  PrecisionAg = "PrecisionAg",
  GPS = "GPS",
  UsgsSoils = "UsgsSoils",
  HighRiskInsured = "HighRiskInsured",
  Insured = "Insured",
  Agency = "Agency",
  FieldPOD = "FieldPOD",
  FieldEvent = "FieldEvent",
  BaseImagery = "BaseImagery",
  USBoundaries = "USBoundaries",
  CLU = "CLU",
  TexasAbstracts = "TexasAbstracts",
  Streets = "Streets",
  PLSS = "PLSS",
  OSM = "OSM",
  TexasCityInfo = "TexasCityInfo",
  RMAImagery = "RMAImagery",
  Layout = "Layout"
}